+++ tests/cpp-tests/Classes/FileUtilsTest/FileUtilsTest.cpp
@@ -373,8 +373,8 @@ void TestFileFuncs::onEnter()
         this->addChild(label);
 
         // getFileSize Test
-        long size = sharedFileUtils->getFileSize(filepath);
-        msg = StringUtils::format("getFileSize: Test file size equals %ld", size);
+        int32_t size = sharedFileUtils->getFileSize(filepath);
+        msg = StringUtils::format("getFileSize: Test file size equals %d", size);
         label = Label::createWithSystemFont(msg, "", 20);
         label->setPosition(x, y * 3);
         this->addChild(label);
@@ -1160,8 +1160,8 @@ void TestFileFuncsAsync::onEnter()
         label->setPosition(x, y * 4);
         this->addChild(label);
         
-        sharedFileUtils->getFileSize(filepath, [=](long size) {
-            auto msg = StringUtils::format("getFileSize: Test file size equals %ld", size);
+        sharedFileUtils->getFileSize(filepath, [=](int32_t size) {
+            auto msg = StringUtils::format("getFileSize: Test file size equals %d", size);
             auto label = Label::createWithSystemFont(msg, "", 20);
             label->setPosition(x, y * 3);
             this->addChild(label);
